Move 3W list inside routed page
